Content negotiations

+++ api/Controllers/MoviesController.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using api.Data;
 using api.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace api.Controllers
@@ -16,47 +17,70 @@ namespace api.Controllers
             _dbContext = dbContext;
         }
 
+
         // GET api/movies
         [HttpGet]
-        public IEnumerable<Movie> Get()
+        public IActionResult Get()
         {
-            return _dbContext.Movies;
+            return Ok(_dbContext.Movies);
         }
+
 
         // GET api/movies/3
         [HttpGet("{id}")]
-        public Movie Get(int id)
+        public IActionResult Get(int id)
         {
             var movie = _dbContext.Movies.Find(id);
-
-            return movie;
+            if(movie == null) 
+            {
+                return NotFound($"Brak rekordu o id {id}");
+            }
+            return Ok(movie);
         }
+
 
         // POST api/movies
         [HttpPost]
-        public void Post([FromBody] Movie movie)
+        public IActionResult Post([FromBody] Movie movie)
         {
             _dbContext.Movies.Add(movie);
             _dbContext.SaveChanges();
+
+            return StatusCode(StatusCodes.Status201Created);
         }
+
 
         // PUT api/movies/3
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Movie movie)
+        public IActionResult Put(int id, [FromBody] Movie movie)
         {
             var movieToPut = _dbContext.Movies.Find(id);
+            if(movieToPut == null) 
+            {
+                return NotFound($"Brak rekordu o id {id}");
+            }
             movieToPut.Name = movie.Name;
             movieToPut.Language = movie.Language;
+            movieToPut.Rating = movie.Rating;
             _dbContext.SaveChanges();
+
+            return Ok($"Zaktualizowano rekord {id}.");
         }
+
 
         // DELETE api/movies/3
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             var movie = _dbContext.Movies.Find(id);
+            if(movie == null) 
+            {
+                return NotFound($"Brak rekordu o id {id}");
+            }
             _dbContext.Movies.Remove(movie);
             _dbContext.SaveChanges();
+
+            return Ok($"Usunięto rekord o id {id}.");
         }
 
     }
